refactor: grouped uuid, me ids, and external ids logic into idutils package
