+++ internal/idutils/me_utils.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package utils
+package idutils
 
 import "regexp"
